+++ src/pages/index.tsx
@@ -20,9 +20,6 @@ import { HeadProps } from 'gatsby';
 import { SEO } from '../components/seo';
 import { DataProps } from '../hooks/use-site-metadata';
 
-import { Footer } from '../theme/components/Footer';
-import Layout from '../theme/components/Layout';
-import { Header } from '../theme/components/Header';
 import {
   Landing,
   ModelGenerationApp,
@@ -30,7 +27,11 @@ import {
   News,
   About,
   Libraries,
-} from '../theme/sections';
+} from '../sections';
+
+import { Footer } from '../theme/components/Footer';
+import Layout from '../theme/components/Layout';
+import { Header } from '../theme/components/Header';
 import { PAGE } from '../theme/utils/constants';
 import { MailingListSubscription } from '../theme/components/mailingListSubscription/MailingListSubscription';
